AIchatbot/requirements.txt
streamlit>=1.32.0
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
openai>=1.40.0
